requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
greenlet>=3.0.0,<4.0.0
playwright==1.47.0
pytest==8.3.3

